Covid_2019_EDA
This dataset is a comprehensive collection of data related to the spread of COVID-19 in India. It captures the number of confirmed cases and deaths in each state and union territory of India.
It is important because it allows us to understand the current situation of the pandemic in India and to monitor the progress of the virus in each state. It can also be used to measure the effectiveness of the strategies implemented by the Indian Government to contain the spread of the virus